Add arlab1 + nuc device launch files
