+++ code/manipulator_description/urdf/manipulator.urdf.xacro
@@ -130,7 +130,7 @@
     name="unten"
     parent_link="world"
     xyz="-0.42 0.42 0"
-    rpy="1.571 0 -0.960">
+    rpy="1.571 0 -0.785">
   </xacro:schrankregal>
 
   <!-- scene -->
@@ -138,7 +138,7 @@
     name="mitte"
     parent_link="world"
     xyz="-0.42 0.42 0.46"
-    rpy="1.571 0 -0.960">
+    rpy="1.571 0 -0.785">
   </xacro:schrankregal>
 
   <!-- scene -->
@@ -146,7 +146,7 @@
     name="oben"
     parent_link="world"
     xyz="-0.42 0.42 0.92"
-    rpy="1.571 0 -0.960">
+    rpy="1.571 0 -0.785">
   </xacro:schrankregal>
 
   <!-- arm -->
